Replaced plot point with entity.
